+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Added
 
-- `size` now supports async iterables
+- `cycle` now supports async iterables
 - `last` now supports async iterables
+- `size` now supports async iterables
 
 ## 0.3.0 - 2024-02-27